refactor: renombrar paquete base de wiki_documental a wiki
