+++ src/wiki/processing/style_map.py
