+++ tests/test_cli.py
@@ -5,8 +5,8 @@ from docx.shared import Pt
 
 from typer.testing import CliRunner
 
-import wiki_documental
-from wiki_documental.cli import app
+import wiki
+from wiki.cli import app
 
 runner = CliRunner()
 
@@ -14,7 +14,7 @@ runner = CliRunner()
 def test_version_option():
     result = runner.invoke(app, ["--version"])
     assert result.exit_code == 0
-    assert wiki_documental.__version__ in result.stdout
+    assert wiki.__version__ in result.stdout
 
 
 def test_full_calls_ensure_pandoc(monkeypatch, tmp_path):
@@ -23,7 +23,7 @@ def test_full_calls_ensure_pandoc(monkeypatch, tmp_path):
     def dummy():
         called["value"] = True
 
-    monkeypatch.setattr("wiki_documental.cli.ensure_pandoc", dummy)
+    monkeypatch.setattr("wiki.cli.ensure_pandoc", dummy)
 
     paths = {
         "originals": tmp_path / "orig",
@@ -50,8 +50,8 @@ def test_full_calls_ensure_pandoc(monkeypatch, tmp_path):
         return Result()
 
     monkeypatch.setattr("subprocess.run", fake_run)
-    monkeypatch.setattr("wiki_documental.processing.docx_to_md.ensure_pandoc", lambda: None)
-    monkeypatch.setattr("wiki_documental.cli.cfg", {"paths": paths, "options": {"cutoff_similarity": 0.5}})
+    monkeypatch.setattr("wiki.processing.docx_to_md.ensure_pandoc", lambda: None)
+    monkeypatch.setattr("wiki.cli.cfg", {"paths": paths, "options": {"cutoff_similarity": 0.5}})
 
     result = runner.invoke(app, ["full"])
     assert result.exit_code == 0
@@ -69,7 +69,7 @@ def test_normalize_command(tmp_path, monkeypatch):
     doc.save(sample)
 
     monkeypatch.setattr(
-        "wiki_documental.cli.cfg", {"paths": {"work": tmp_path}}
+        "wiki.cli.cfg", {"paths": {"work": tmp_path}}
     )
     result = runner.invoke(app, ["normalize", str(sample)])
     assert result.exit_code == 0
@@ -82,7 +82,7 @@ def test_map_command(tmp_path, monkeypatch):
     md_folder = tmp_path / "md_raw"
     md_folder.mkdir()
     (md_folder / "sample.md").write_text("# Title\n", encoding="utf-8")
-    monkeypatch.setattr("wiki_documental.cli.cfg", {"paths": {"work": tmp_path}})
+    monkeypatch.setattr("wiki.cli.cfg", {"paths": {"work": tmp_path}})
     result = runner.invoke(app, ["map"])
     assert result.exit_code == 0
     map_file = tmp_path / "map.yaml"
@@ -101,7 +101,7 @@ def test_index_overwrite(tmp_path, monkeypatch):
     work = tmp_path
     (work / "map.yaml").write_text(yaml.safe_dump(map_data, allow_unicode=True), encoding="utf-8")
     (work / "index.yaml").write_text("old", encoding="utf-8")
-    monkeypatch.setattr("wiki_documental.cli.cfg", {"paths": {"work": work}})
+    monkeypatch.setattr("wiki.cli.cfg", {"paths": {"work": work}})
 
     result = runner.invoke(app, ["index", "--overwrite"])
     assert result.exit_code == 0
@@ -117,7 +117,7 @@ def test_sidebar_command(tmp_path, monkeypatch):
     work = tmp_path
     (work / "index.yaml").write_text(yaml.safe_dump(index, allow_unicode=True), encoding="utf-8")
     paths = {"work": work, "wiki": work}
-    monkeypatch.setattr("wiki_documental.cli.cfg", {"paths": paths})
+    monkeypatch.setattr("wiki.cli.cfg", {"paths": paths})
     (work / "README.md").write_text("intro", encoding="utf-8")
 
     result = runner.invoke(app, ["sidebar"])
@@ -147,7 +147,7 @@ def test_sidebar_command_depth(tmp_path, monkeypatch):
     work = tmp_path
     (work / "index.yaml").write_text(yaml.safe_dump(index, allow_unicode=True), encoding="utf-8")
     paths = {"work": work, "wiki": work}
-    monkeypatch.setattr("wiki_documental.cli.cfg", {"paths": paths})
+    monkeypatch.setattr("wiki.cli.cfg", {"paths": paths})
     (work / "README.md").write_text("intro", encoding="utf-8")
 
     result = runner.invoke(app, ["sidebar", "--depth", "2"])
