+++ tests/test_docx_to_md.py
@@ -1,6 +1,6 @@
 from pathlib import Path
 from docx import Document
-from wiki_documental.processing.docx_to_md import convert_docx_to_md
+from wiki.processing.docx_to_md import convert_docx_to_md
 
 
 def _create_sample_docx(path: Path, text: str) -> None:
@@ -23,7 +23,7 @@ def test_convert_docx_to_md(tmp_path, monkeypatch):
 
     monkeypatch.setattr("subprocess.run", fake_run)
     monkeypatch.setattr(
-        "wiki_documental.processing.docx_to_md.ensure_pandoc", lambda: None
+        "wiki.processing.docx_to_md.ensure_pandoc", lambda: None
     )
     convert_docx_to_md(docx_file, md_file)
     assert md_file.exists()
