+++ tests/test_reset_work_dir.py
@@ -3,7 +3,7 @@ from docx import Document
 from docx.shared import Pt
 from typer.testing import CliRunner
 
-from wiki_documental.cli import app
+from wiki.cli import app
 
 runner = CliRunner()
 
@@ -42,9 +42,9 @@ def test_reset_work_dir(tmp_path, monkeypatch):
     _create_doc(doc_file)
 
     monkeypatch.setattr("subprocess.run", _fake_run)
-    monkeypatch.setattr("wiki_documental.processing.docx_to_md.ensure_pandoc", lambda: None)
-    monkeypatch.setattr("wiki_documental.cli.ensure_pandoc", lambda: None)
-    monkeypatch.setattr("wiki_documental.cli.cfg", {"paths": paths, "options": {"cutoff_similarity": 0.5}})
+    monkeypatch.setattr("wiki.processing.docx_to_md.ensure_pandoc", lambda: None)
+    monkeypatch.setattr("wiki.cli.ensure_pandoc", lambda: None)
+    monkeypatch.setattr("wiki.cli.cfg", {"paths": paths, "options": {"cutoff_similarity": 0.5}})
 
     result = runner.invoke(app, ["full"])
     assert result.exit_code == 0
